third_party/WebKit/LayoutTests/inspector/console/console-tainted-globals-expected.txt

Tests that overriding global methods (like Array.prototype.push, Math.max) will not break the inspector.


Running: evaluateInConsole

Running: testRuntimeAgentCallFunctionOn

Running: dumpConsoleMessages
testOverriddenArrayPushAndMathMax()
[1, 2, 3]
testOverriddenConstructorName()
Object {constructor: Object}
testThrowConstructorName()
Object {}
testOverriddenIsFinite()
["arg1", "arg2"]
testOverriddenError()
Object {result: "PASS"}
restoreError()
Object {result: "PASS"}
testOverriddenToString(function func() {}, true)
#<Function>
testOverriddenToString(function func() {}, false)
#<Function>
testOverriddenToString(new Function, true)
#<Function>
testOverriddenToString(new Function, false)
#<Function>
testOverriddenToString(/^regex$/, true)
#<RegExp>
testOverriddenToString(/^regex$/, false)
#<RegExp>
testOverriddenToString(new Date, true)
#<Date>
testOverriddenToString(new Date, false)
#<Date>
testOverriddenToString({}, true)
Object {}
testOverriddenToString({}, false)
Object {}
testOverriddenToString(new Number(1), true)
Number {[[PrimitiveValue]]: 1}
testOverriddenToString(new Number(1), false)
Number {[[PrimitiveValue]]: 1}
